+++ modules.tf
@@ -47,7 +47,7 @@ module "logs" {
 
 module "keyvault" {
   source  = "claranet/keyvault/azurerm"
-  version = "5.1.0"
+  version = "6.0.0"
 
   client_name         = var.client_name
   environment         = var.environment