+++ modules/monitoring_function/r-function.tf
@@ -1,6 +1,6 @@
 module "function" {
   source  = "claranet/function-app/azurerm"
-  version = "6.2.1"
+  version = "7.0.0"
 
   client_name         = var.client_name
   environment         = var.environment
@@ -24,10 +24,14 @@ module "function" {
     SFX_EXTRA_DIMENSIONS         = local.extra_dimensions
   }, var.extra_application_settings)
 
-  os_type                     = "Linux"
-  sku_name                    = "Y1"
-  function_language_for_linux = "python"
-  function_app_version        = 3
+  os_type              = "Linux"
+  sku_name             = "Y1"
+  function_app_version = 3
+  function_app_site_config = {
+    application_stack = {
+      python_version = "3.8"
+    }
+  }
 
   # Allow the current run to access Storage Table for FAME queries
   storage_account_network_rules_enabled = var.storage_account_network_rules_enabled