+++ versions.tf
@@ -1,9 +1,9 @@
 terraform {
-  required_version = ">= 1.1, < 1.3"
+  required_version = ">= 1.3"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.1"
+      version = "~> 3.22"
     }
     # tflint-ignore: terraform_unused_required_providers
     azurecaf = {