+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Changed
   * AZ-800: Automation account can only have 15 tags maximum
+  * AZ-846: Bump diagnotics settings module
 
 # v6.2.0 - 2022-09-09